Create HumanYears.swift
Hey there!  This is a variation of the DogAge.swift program that we learned.  I thought it would be cool to flip it around and see how old I was in dog years so I reversed the math that I saw in the DogYears program and came up with this!  Plus I got a lot of good practice with commenting, data type declaration, variable manipulation and understanding, and the "let" and "var" keywords.  I hope you like it, and if there is anything about it that makes you unable to upload it, please ~let~ me know and I'll be happy to edit it.  I'd love to have this up here for people to see :)!!

Thanks again.
Stay safe, healthy, and happy

Sincerely,
Jacob Kline