Add challenge solutions
Added solutions for the 3 optional challenges at the end of the project.

Also included example of how to use a for-loop to solve challenge 1.
I felt this was important because most Codecademy learners -- even
those who know other languages or have completed the entire
Learn Swift course -- would have trouble using a for-loop for
Challenge #1 due to the nature of Swift's sets.

 Sets in Swift are value types and cannot be subscripted by their index
 the way arrays can. This presents unique issues when trying to loop
 through a set of sets to perform manipulations on those sets.