I believe this removes the limits of what integer may be picked for var checkPrime. This also allows var isPrime to be set to either true or false from initialization.
In the original code, there is a comment on line 15 that contains a small mistake
The code has an error when checkPrime = 2, because a range cannot have the upper and lower bound be the same value. This can occur because the range will be 2...(2-1).
maybe this works in swift, but it didn't work on Codeacademy for some reason.